+++ addon/components/paper-autocomplete-content.js
@@ -1,0 +1,3 @@
+import ContentComponent from 'ember-basic-dropdown/components/basic-dropdown/content';
+
+export default ContentComponent;