+++ app/components/base-focusable.js
@@ -1,3 +1,0 @@
-import BaseFocusable from 'ember-paper/components/base-focusable';
-
-export default BaseFocusable;