+++ app/components/paper-virtual-repeat-scroller.js
@@ -1,0 +1,3 @@
+import PaperVirtualRepeatScroller from 'ember-paper/components/paper-virtual-repeat-scroller';
+
+export default PaperVirtualRepeatScroller;