+++ tests/dummy/app/controllers/demo/select.js
@@ -1,27 +1,15 @@
 import Ember from 'ember';
 
-const { computed, Controller, A, RSVP, run } = Ember;
+const { Controller, computed, RSVP, A, run } = Ember;
 
 export default Controller.extend({
 
   userState: '',
   states: computed(function() {
-    return 'AL AK AZ AR CA CO CT DE FL GA HI ID IL IN IA KS KY LA ME MD MA MI MN MS MO MT NE NV NH NJ NM NY NC ND OH OK OR PA RI SC SD TN TX UT VT VA WA WV WI WY'
-            .split(' ').map((state) => ({ abbrev: state }));
+    return A('AL AK AZ AR CA CO CT DE FL GA HI ID IL IN IA KS KY LA ME MD MA MI MN MS MO MT NE NV NH NJ NM NY NC ND OH OK OR PA RI SC SD TN TX UT VT VA WA WV WI WY'
+            .split(' ').map((state) => ({ abbrev: state })));
   }),
-  vegetables: A([
-    { name: 'Corn', checked: false },
-    { name: 'Onions', checked: false },
-    { name: 'Kale', checked: false },
-    { name: 'Arugula', checked: false },
-    { name: 'Peas', checked: false },
-    { name: 'Zucchini', checked: false }]),
-  searchTerm: '',
-  filteredVegetables: computed('vegetables.[]', 'searchTerm', function() {
-    return this.get('vegetables').filter((vegetable) => {
-      return vegetable.name.toLowerCase().indexOf(this.get('searchTerm').toLowerCase()) > -1;
-    });
-  }),
+
   sizes: A([
     'small (12-inch)',
     'medium (14-inch)',
@@ -29,10 +17,18 @@ export default Controller.extend({
     'insane (42-inch)'
   ]),
 
+  vegetables: A([
+    { name: 'Corn', checked: false },
+    { name: 'Onions', checked: false },
+    { name: 'Kale', checked: false },
+    { name: 'Arugula', checked: false },
+    { name: 'Peas', checked: false },
+    { name: 'Zucchini', checked: false }
+  ]),
   /*
    * Fake promise to fetch data, here you would use ember-data, jQuery.ajax or whatever you want.
    */
-  onUserLoad() {
+  users: computed(function() {
     let _self = this;
     return new RSVP.Promise(function(resolve) {
       // Just wait for 800ms to 2 seconds for a fake progress, so it feels like a query.
@@ -56,33 +52,9 @@ export default Controller.extend({
       }, waitMS);
 
     });
-  },
-
-  userLabelCallback(item) {
-    // using ember data, this might be 'item.get('name')'
-    return item.name;
-  },
-  vegetableLabelCallback(item) {
-    return item.name;
-  },
-  toppings: A([
-    { category: 'meat', name: 'Pepperoni' },
-    { category: 'meat', name: 'Sausage' },
-    { category: 'meat', name: 'Ground Beef' },
-    { category: 'meat', name: 'Bacon' },
-    { category: 'veg', name: 'Mushrooms' },
-    { category: 'veg', name: 'Onion' },
-    { category: 'veg', name: 'Green Pepper' },
-    { category: 'veg', name: 'Green Olives' }
-  ]),
-
-  meatToppings: computed.filterBy('toppings', 'category', 'meat'),
-
-  vegToppings: computed.filterBy('toppings', 'category', 'veg'),
-  actions: {
-    searchKeyPressed(e) {
-      e.stopPropagation();
-    }
-  }
-
+  }),
+  groupedToppings: [
+    { groupName: 'Meats', options: ['Pepperoni', 'Sausage', 'Ground Beef', 'Bacon'] },
+    { groupName: 'Veg', options: ['Mushrooms', 'Onion', 'Green Pepper', 'Green Olives'] }
+  ]
 });